Moves handler and proxy cliargs into respective folders
