+++ src/core/mod.rs
@@ -23,14 +23,12 @@ pub mod event;
 pub mod eventloop;
 pub mod executor;
 pub mod fs;
-pub mod handlercliargs;
 pub mod http1;
 pub mod io;
 pub mod jwt;
 pub mod list;
 pub mod log;
 pub mod net;
-pub mod proxycliargs;
 pub mod reactor;
 pub mod select;
 pub mod shuffle;
